# src/main/resources/static/booking/booking-openapi.yaml
openapi: '3.0.0'
info:
  title: Booking API
  description: Booking API
  version: 1.0.0
servers:
  - url: http://localhost:8080
...
paths:
...
  /shows/{show_id}:
    get:
      summary: Get a show
      operationId: getShow
      tags:
        - Show
      parameters:
        - in: path
          name: show_id
          description: Show ID
          required: true
          schema:
            $ref: '#/components/schemas/ShowId'
      responses:
        200:
          description: Get show response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetShowResponse'
        default:
          $ref: '#/components/responses/Problem'
  /shows:
    get:
      summary: List all shows
      operationId: listShows
      tags:
        - Show
      parameters:
        - in: query
          name: offset
          description: Offset
          schema:
            type: integer
            format: int64
            minimum: 0
            default: 0
        - in: query
          name: limit
          description: Limit
          schema:
            type: integer
            format: int32
            minimum: 1
            maximum: 100
            default: 10
      responses:
        200:
          description: List shows response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListShowsResponse'
        default:
          $ref: '#/components/responses/Problem'
  /search/shows:
    get:
      summary: Search all shows
      operationId: searchShows
      tags:
        - Show
      parameters:
        - in: query
          name: q
          description: Query
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 20
        - in: query
          name: offset
          description: Offset
          schema:
            type: integer
            format: int64
            minimum: 0
            default: 0
        - in: query
          name: limit
          description: Limit
          schema:
            type: integer
            format: int32
            minimum: 1
            maximum: 100
            default: 10
      responses:
        200:
          description: Search shows response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchShowsResponse'
        default:
          $ref: '#/components/responses/Problem'
components:
  schemas:
...
    GetShowResponse:
      type: object
      description: Get show response
      required:
        - showId
        - scheduledAt
        - movie
        - hall
        - seats
      properties:
        showId:
          $ref: '#/components/schemas/ShowId'
        scheduledAt:
          type: string
          format: date-time
        movie:
          $ref: '#/components/schemas/ShowMovie'
        hall:
          $ref: '#/components/schemas/ShowHall'
        seats:
          items:
            $ref: '#/components/schemas/Seat'
    Seat:
      type: object
      description: Show seat
      required:
        - seatNumber
        - status
      properties:
        seatNumber:
          $ref: '#/components/schemas/SeatNumber'
        status:
          $ref: '#/components/schemas/SeatStatus'
        booking:
          $ref: '#/components/schemas/SeatBooking'
    SeatBooking:
      type: object
      description: Seat booking
      required:
        - bookingId
        - status
      properties:
        bookingId:
          $ref: '#/components/schemas/BookingId'
        status:
          $ref: '#/components/schemas/BookingStatus'
    ListShowsResponse:
      type: object
      description: List shows response
      required:
        - shows
      properties:
        shows:
          items:
            $ref: '#/components/schemas/ShowSummary'
    SearchShowsResponse:
      type: object
      description: Search shows response
      required:
        - shows
      properties:
        shows:
          items:
            $ref: '#/components/schemas/ShowSummary'
    ShowSummary:
      type: object
      description: Show summary
      required:
        - showId
        - scheduledAt
        - movie
        - hall
      properties:
        showId:
          $ref: '#/components/schemas/ShowId'
        scheduledAt:
          type: string
          format: date-time
        movie:
          $ref: '#/components/schemas/ShowMovie'
        hall:
          $ref: '#/components/schemas/ShowHall'
    ShowMovie:
      type: object
      description: Show movie
      required:
        - movieId
        - title
      properties:
        movieId:
          $ref: '#/components/schemas/MovieId'
        title:
          type: string
    ShowHall:
      type: object
      description: Show hall
      required:
        - hallId
        - name
      properties:
        hallId:
          $ref: '#/components/schemas/HallId'
        name:
          type: string
    BookingId:
      type: string
      pattern: '[A-Za-z0-9_-]{11}'
      description: Booking ID
    BookingStatus:
      type: string
      description: Booking status
      x-extensible-enum:
        - INITIATED
        - CONFIRMED
        - CANCELLED
    HallId:
      type: string
      pattern: '[A-Za-z0-9_-]{11}'
      description: Hall ID
    MovieId:
      type: string
      pattern: '[A-Za-z0-9_-]{11}'
      description: Movie ID
    ShowId:
      type: string
      pattern: '[A-Za-z0-9_-]{11}'
      description: Show ID
    SeatNumber:
      type: string
      pattern: '[A-Z][1-9][0-9]?'
      description: Seat number
      example: 'A1'
    SeatStatus:
      type: string
      description: Seat status
      x-extensible-enum:
        - AVAILABLE
        - RESERVED
        - BOOKED
    ProblemResponse:
      type: object
      description: Problem response
      required:
        - type
        - title
        - status
        - instance
      properties:
        type:
          type: string
          format: uri
          example: urn:not-found
        title:
          type: string
          example: Not Found
        status:
          type: integer
          format: int32
          minimum: 100
          maximum: 600
          exclusiveMaximum: true
          example: 404
        detail:
          type: string
          example: Resource not found
        instance:
          type: string
          format: uri
          example: /resource
  responses:
    Problem:
      description: Problem
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemResponse'
...